Modified movement and objects
Fixed some errors that were found during playtesting such as the cat flipping on itself (fixed by locking the Z-axis on the cat). Modified the Movement script to use force impulses instead of instantly changing the velocity. Added variety  in objects. Made prefabs of all the objects instead of just having them out of frame to the side. Made it so that the rocks disappear when they land on good objects(done with tags, currently hardcoded).